renaming x2
