+++ skbio/stats/ordination/_redundancy_analysis.py
@@ -12,7 +12,7 @@ from scipy.linalg import svd, lstsq
 
 from ._ordination_results import OrdinationResults
 from ._utils import corr, svd_rank, scale
-from skbio.table._dispatcher import _ingest_table, _create_table, _create_table_1d
+from skbio.table._tabular import _ingest_table, _create_table, _create_table_1d
 
 
 def rda(
